(ColorField) Renames from FormFieldColorPicker to FormFieldColor
