+++ src/components/Field/FormFieldColor/FormFieldColor.stories.mdx
@@ -7,7 +7,7 @@ This field implements the [**FieldDef**](#generic-field-props-fielddef) interfac
 
 - The color picker component allows users to select from pre-defined colors (swatches) or custom colors using a HSB selection interface.
 - The implementation of this component is based on the following package: [React Color](https://casesandberg.github.io/react-color/)
-- [**Playground**](/?path=/story/formfields-formfieldcolorpicker--playground)
+- [**Playground**](/?path=/story/formfields-formfieldcolor--playground)
 - Data: `string` - Text that represents the hex value of the color picker.
 - inputSettings: This component doesn't require any additional props.
