+++ src/components/Field/FormFieldColor/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./FormFieldColor";
+export * from "./FormFieldColor";
