--- src/components/Field/FormFieldColorPicker/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./FormFieldColorPicker";
-export * from "./FormFieldColorPicker";
