+++ src/components/Field/index.tsx
@@ -13,8 +13,8 @@ export * from "./FormFieldCheckbox";
 export { default as FormFieldChips } from "./FormFieldChips";
 export * from "./FormFieldChips";
 
-export { default as FormFieldColorPicker } from "./FormFieldColorPicker";
-export * from "./FormFieldColorPicker";
+export { default as FormFieldColorPicker } from "./FormFieldColor";
+export * from "./FormFieldColor";
 
 export { default as FormFieldDate } from "./FormFieldDate/DateField";
 export * from "./FormFieldDate/DateField";
